Fix: /dev/null permission denied errors during the build
Creating `/dev/null` while the root file system initialization for some reason
seem render it unusable for the commands running in chrooted environment.

Not creating `/dev/null` seem to get rid of the errors, except for installing
the package `gnupg`.

Will have to look into this later if this is causing a problem.